services/main-service/requirements.txt
Flask
Flask-SQLAlchemy
SQLAlchemy
Werkzeug
psycopg2-binary
Flask-Migrate
pyjwt
cryptography
